% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_methods_md}
\alias{set_methods_md}
\title{Create methods node from .md file}
\usage{
set_methods_md(f)
}
\arguments{
\item{f}{(character) Full path to methods.md file}
}
\value{
(list) EML methodStep node for markdown as supported by the Environmental Data Initiative Data Repository
}
\description{
Create methods node from .md file
}
